# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: GustoPOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drinks:
    get:
      operationId: getDrinks
      tags: [drinks]
      summary: List all drinks on the menu
      responses:
        "200":
          description: List of drinks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Drink"
    post:
      operationId: createDrink
      tags: [drinks]
      summary: Create a drink with recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDrinkBody"
      responses:
        "201":
          description: Created drink
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drink"
  /drinks/{id}:
    get:
      operationId: getDrink
      tags: [drinks]
      summary: Get a drink
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Drink
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drink"
    patch:
      operationId: updateDrink
      tags: [drinks]
      summary: Update a drink
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDrinkBody"
      responses:
        "200":
          description: Updated drink
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drink"
    delete:
      operationId: deleteDrink
      tags: [drinks]
      summary: Delete a drink
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/backups/{id}/restore:
    post:
      operationId: restoreBackup
      tags: [admin]
      summary: Restore from a backup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backup restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/backups/{id}:
    delete:
      operationId: deleteBackup
      tags: [admin]
      summary: Delete a backup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backup deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    RecipeIngredient:
      type: object
      properties:
        ingredientId:
          type: ["string", "null"]
        ingredientName:
          type: string
        ingredientNameEs:
          type: ["string", "null"]
        amountInBaseUnit:
          type: number
        costContribution:
          type: number
      required:
        - ingredientName
        - amountInBaseUnit
        - costContribution
    Drink:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameEs:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        descriptionEs:
          type: ["string", "null"]
        category:
          type: string
        sourceType:
          type: string
          default: standard
        costPerDrink:
          type: number
        suggestedPrice:
          type: number
        actualPrice:
          type: ["number", "null"]
        markupFactor:
          type: number
        recipe:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        isAvailable:
          type: boolean
        isOnMenu:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - costPerDrink
        - suggestedPrice
        - actualPrice
        - markupFactor
        - recipe
        - isAvailable
        - isOnMenu
        - createdAt
        - updatedAt
    CreateDrinkBody:
      type: object
      properties:
        name:
          type: string
        nameEs:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        descriptionEs:
          type: ["string", "null"]
        category:
          type: string
        markupFactor:
          type: number
        actualPrice:
          type: ["number", "null"]
        isAvailable:
          type: boolean
        recipe:
          type: array
          items:
            $ref: "#/components/schemas/CreateRecipeIngredientBody"
      required:
        - name
        - category
    CreateRecipeIngredientBody:
      type: object
      properties:
        ingredientId:
          type: string
        amountInBaseUnit:
          type: number
      required:
        - ingredientId
        - amountInBaseUnit
    UpdateDrinkBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        nameEs:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        descriptionEs:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        markupFactor:
          type: ["number", "null"]
        actualPrice:
          type: ["number", "null"]
        isAvailable:
          type: ["boolean", "null"]
        recipe:
          type: array
          items:
            $ref: "#/components/schemas/CreateRecipeIngredientBody"
...